Moved some Make stuff from subdirs to shared Makefiles.
