+++ src/Makefile.am
@@ -1,4 +1,5 @@
 SUBDIRS =
+pkginclude_HEADERS =
 lib_LTLIBRARIES = libopenscap.la
 noinst_LTLIBRARIES = libopenscap_testing.la
 
@@ -10,7 +11,50 @@ libopenscap_testing_la_SOURCES =
 libopenscap_testing_la_LIBADD = $(top_builddir)/lib/libgnu.la
 libopenscap_testing_la_LDFLAGS= -rpath /nowhere
 
-SUBDIRS += common
+noinst_LTLIBRARIES += liboscapcommon.la
+
+liboscapcommon_la_SOURCES = \
+	common/alloc.c common/alloc.h \
+	common/assume.h \
+	common/bfind.c common/bfind.h \
+	common/debug.c common/debug_priv.h \
+	common/elements.c common/elements.h \
+	common/err_queue.c common/err_queue.h \
+	common/error.c common/_error.h \
+	common/list.c common/list.h \
+	common/memusage.c common/memusage.h \
+	common/oscap_acquire.c common/oscap_acquire.h \
+	common/oscapxml.c common/oscapxml.h \
+	common/oscap_buffer.c common/oscap_buffer.h \
+	common/oscap_string.c common/oscap_string.h \
+	common/reference.c common/reference_priv.h \
+	common/text.c common/text_priv.h \
+	common/tsort.c common/tsort.h \
+	common/util.c common/util.h \
+	common/xml_iterate.c common/xml_iterate.h \
+	common/xmlns_priv.h \
+	common/xmltext_priv.c common/xmltext_priv.h
+
+liboscapcommon_la_CPPFLAGS  = \
+	@curl_CFLAGS@ \
+	@xml2_CFLAGS@ @xslt_CFLAGS@ @exslt_CFLAGS@ \
+	-I$(srcdir)/common/public \
+	-I$(top_srcdir)/src \
+	-I$(top_srcdir)/src/OVAL/probes/SEAP/public \
+	-I$(top_srcdir)/src/source/public
+
+liboscapcommon_la_LIBADD = \
+	@curl_LIBS@ \
+	@xml2_LIBS@ @xslt_LIBS@ @exslt_LIBS@ @PTHREAD_LIBS@
+
+pkginclude_HEADERS += \
+	common/public/oscap_debug.h \
+	common/public/oscap_error.h \
+	common/public/oscap.h \
+	common/public/oscap_reference.h \
+	common/public/oscap_download_cb.h \
+	common/public/oscap_text.h
+
 libopenscap_la_LIBADD += common/liboscapcommon.la
 libopenscap_testing_la_LIBADD += common/liboscapcommon.la
