+++ tests/API/XCCDF/unittests/test_xccdf_overrides.c
@@ -28,7 +28,7 @@
 #include <xccdf_policy.h>
 
 #include "unit_helper.h"
-#include <../../../assume.h>
+#include "../../assume.h"
 
 int main(int argc, char *argv[])
 {
