+++ tests/API/XCCDF/variable_instance/Makefile.am
@@ -1,16 +1,9 @@
-DISTCLEANFILES = *.log *.out* oscap_debug.log.*
-CLEANFILES = *.log *.out* oscap_debug.log.*
 
-TESTS_ENVIRONMENT = \
-		builddir=$(top_builddir) \
-		OSCAP_FULL_VALIDATION=1 \
-		$(top_builddir)/run
+TESTS += %reldir%/all.sh
 
-TESTS = all.sh
-
-EXTRA_DIST = \
-	all.sh \
-	requires_both-oval.xml \
-	testing_file_300.xml \
-	testing_file_600.xml \
-	test_xccdf_variable_instance.xccdf.xml
+EXTRA_DIST += \
+	%reldir%/all.sh \
+	%reldir%/requires_both-oval.xml \
+	%reldir%/testing_file_300.xml \
+	%reldir%/testing_file_600.xml \
+	%reldir%/test_xccdf_variable_instance.xccdf.xml
